requirements.txt
psycopg2-binary>=2.9.9
openai>=1.12.0
python-dotenv>=1.0.0
langchain>=0.1.0
langchain-openai>=0.0.5
groq>=0.4.0
streamlit>=1.30.0
pandas>=2.0.0
sqlparse>=0.4.4
pytest>=7.0.0
pytest-cov>=4.0.0
